Setup a test to find where the NaN comes from

+++ src/modules/ImuCamCalibration.cpp
@@ -34,7 +34,27 @@ public:
 	auto imuInput = inputs.getIMUInput("imu");
 	if (auto imuData = imuInput.data()) {
 	  for (const auto &singleImu : imuData) {
-		calibrator.addImu(singleImu.timestamp, singleImu.gyroscopeX, singleImu.gyroscopeY, singleImu.gyroscopeZ, singleImu.accelerometerX, singleImu.accelerometerY, singleImu.accelerometerZ);
+
+
+	// TODO(radam): del
+//		std::stringstream ss;
+//		ss << "/tmp/imu_" << singleImu.timestamp << ".txt";
+//		std::ofstream myfile(ss.str());
+//		myfile << singleImu.timestamp << std::endl;
+//		myfile << singleImu.gyroscopeX << std::endl;
+//		myfile <<  singleImu.gyroscopeY<< std::endl;
+//		myfile <<  singleImu.gyroscopeZ<< std::endl;
+//		myfile <<  singleImu.accelerometerX<< std::endl;
+//		myfile <<  singleImu.accelerometerY<< std::endl;
+//		myfile <<  singleImu.accelerometerZ << std::endl;
+
+		calibrator.addImu(singleImu.timestamp,
+					static_cast<double>(singleImu.gyroscopeX),
+						static_cast<double>(singleImu.gyroscopeY),
+							static_cast<double>(singleImu.gyroscopeZ),
+								static_cast<double>(singleImu.accelerometerX),
+									static_cast<double>(singleImu.accelerometerY),
+										static_cast<double>(singleImu.accelerometerZ));
 	  }
 	}
 
@@ -43,6 +63,13 @@ public:
 	if (auto frame = frameInput.data()) {
 	  cv::Mat img = *frame.getMatPointer();
 	  calibrator.addImage(img, frame.timestamp());
+
+
+	  // TODO(radam): del
+//
+//	  std::stringstream ss;
+//	  ss << "/tmp/" << frame.timestamp() << ".png";
+//	  cv::imwrite(ss.str(), img);
 	}
 
 	// Output preview image
